pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = "personally customized bot with langchain and gpt4"
authors = ["minwook <minkj1992@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
langchain = "^0.0.141"
faiss-cpu = "^1.7.3"
bs4 = "^0.0.1"
xmltodict = "^0.13.0"
openai = "^0.27.4"
requests = "^2.28.2"
tiktoken = "^0.3.3"
python-dotenv = "^1.0.0"
transformers = "^4.28.1"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
aioredis = "^2.0.1"
redis = "^4.5.4"
redis-om = "^0.1.2"
websockets = "^11.0.2"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
